Prove WeakPNT(_real) and WeakPNT'
I realised that the WeakPNT Lean definition takes a limit over integers,
which *I believe* is usually not what one needs in Lean, especially in
number theory. So I renamed the previous WeakPNT to WeakPNT_nat, and
stated + proved WeakPNT where the domain is the reals.

Also proved WeakPNT', which is the statement written down in the LaTeX /
blueprint, and relabelled the blueprint.

(I need this for some further work in Corollaries, not just a random
change.)